Poll.cpp added
